Put OwnershipData<> in its own header
