+++ code/krEngine/ownership.h
@@ -1,4 +1,5 @@
 #pragma once
 
+#include <krEngine/ownership/ownershipData.h>
 #include <krEngine/ownership/owned.h>
 #include <krEngine/ownership/borrowed.h>
