+++ code/krEngine/ownership/owned.h
@@ -1,23 +1,10 @@
 #pragma once
 
-#include <Foundation/Threading/AtomicUtils.h>
 #include <krEngine/common/traits.h>
+#include <krEngine/ownership/ownershipData.h>
 
 namespace kr
 {
-  template<typename T>
-  struct OwnershipData
-  {
-    T* ptr = nullptr;
-    mutable ezInt64 refCount = 0;
-  };
-
-  template<typename T>
-  void addRef(const OwnershipData<T>& data) { ezAtomicUtils::Increment(data.refCount); }
-
-  template<typename T>
-  void removeRef(const OwnershipData<T>& data) { ezAtomicUtils::Decrement(data.refCount); }
-
   /// \brief Wrapper for ownership over an instance of an object.
   template<typename T>
   struct Owned
